Implement the MatchDocument method
In the first tuple element, return all the plus words of the query contained in the document. The words may not be duplicated. Sort them in ascending order.
If the document does not match the query (there are no plus or minus words), the word vector should be returned empty.